requirements.txt
bilibili-api-python==17.1.2
tenacity>=8.0.0
requests>=2.25.0
aiohttp>=3.8.0
openai-whisper>=20230918
tqdm>=4.60.0
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
python-multipart>=0.0.6
starlette>=0.27.0
email-validator>=2.0.0
# Added for V2 vision+LLM pipeline
httpx>=0.27.0
pillow>=10.0.0
numpy>=1.23.0
opencv-python>=4.8.0
pydantic-settings>=2.2.1
rich>=13.7.0
